mainproject/requirements.txt
appdirs==1.4.3
apturl==0.5.2
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
catfish==1.4.13
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.13.1
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
menulibre==2.2.1
monotonic==1.5
more-itertools==4.2.0
mugshot==0.4.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==2023.11.15
platformdirs==3.11.0
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sgt-launcher==0.2.5
simplejson==3.16.0
sip==4.19.21
six==1.14.0
system-service==0.3
systemd-python==234
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
xcffib==0.8.1
xkit==0.0.0
zipp==1.0.0
